refactor: added mock response plugin
orgainized proxy-plugins project
added namespace to abstractons and plugin projects
converted console.write calls to use an ILogger
added common response headers uitl to ProxyUtils
