+++ msgraph-developer-proxy-abstractions/ResponseState.cs
@@ -1,5 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
+
+namespace Microsoft.Graph.DeveloperProxy.Abstractions;
+
 public class ResponseState {
     public bool HasBeenSet { get; set; } = false;
 }
