--- msgraph-developer-proxy/ErrorResponseBody.cs
@@ -1,8 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using System.Text.Json.Serialization;
-
-namespace Microsoft.Graph.DeveloperProxy {
-
-}
