+++ msgraph-developer-proxy/Program.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using Microsoft.Graph.DeveloperProxy;
+using Microsoft.Graph.DeveloperProxy.Abstractions;
 using System.CommandLine;
 
 ILogger logger = new ConsoleLogger();
@@ -9,7 +10,7 @@ IProxyContext context = new ProxyContext(logger);
 ProxyHost proxyHost = new();
 RootCommand rootCommand = proxyHost.GetRootCommand();
 PluginEvents pluginEvents = new PluginEvents();
-PluginLoaderResult loaderResults = new PluginLoader().LoadPlugins(pluginEvents, context);
+PluginLoaderResult loaderResults = new PluginLoader(logger).LoadPlugins(pluginEvents, context);
 
 // have all the plugins init and provide any command line options
 pluginEvents.FireInit(new InitArgs(rootCommand));
