+++ msgraph-developer-proxy/appsettings.json
@@ -1,8 +1,22 @@
 {
   "handlers": [
     {
+      "pluginPath": "GraphProxyPlugins\\msgraph-proxy-plugins.dll",
+      "name": "SelectGuidancePlugin"
+    },
+    {
+      "pluginPath": "GraphProxyPlugins\\msgraph-proxy-plugins.dll",
+      "name": "SdkGuidancePlugin"
+    },
+    {
+      "configSection": "mocksPlugin",
+      "pluginPath": "GraphProxyPlugins\\msgraph-proxy-plugins.dll",
+      "name": "MockResponsePlugin",
+      "disabled": false
+    },
+    {
       "configSection": "randomErrorsPlugin",
-      "handlerPath": "GraphProxyPlugins\\msgraph-proxy-plugins.dll",
+      "pluginPath": "GraphProxyPlugins\\msgraph-proxy-plugins.dll",
       "name": "RandomErrorPlugin",
       "urlsToWatch": [
         "https://graph.microsoft.com/v1.0/*",
@@ -13,15 +27,8 @@
         "https://dod-graph.microsoft.us/beta/*",
         "https://microsoftgraph.chinacloudapi.cn/v1.0/*",
         "https://microsoftgraph.chinacloudapi.cn/beta/*"
-      ]
-    },
-    {
-      "handlerPath": "GraphProxyPlugins\\msgraph-proxy-plugins.dll",
-      "name": "SelectGuidancePlugin"
-    },
-    {
-      "handlerPath": "GraphProxyPlugins\\msgraph-proxy-plugins.dll",
-      "name": "SdkGuidancePlugin"
+      ],
+      "disabled":  false
     }
   ],
   "urlsToWatch": [
@@ -38,6 +45,9 @@
     "https://*.sharepoint-df.*/*_api/*",
     "https://*.sharepoint-df.*/*_vti_bin/*"
   ],
+  "mocksPlugin": {
+    "mocksFile":  "mocks.json"
+  },
   "randomErrorsPlugin": {
     "rate": 99,
     "allowedErrors": [ 429, 500, 502, 503, 504, 507 ]
