+++ msgraph-proxy-plugins/msgraph-proxy-plugins.csproj
@@ -6,13 +6,14 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
 	<EnableDynamicLoading>true</EnableDynamicLoading>
+	<RunPostBuildEvent>OnOutputUpdated</RunPostBuildEvent>
   </PropertyGroup>
   <ItemGroup>
-	  <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0" >
+	  <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0">
 		  <Private>false</Private>
 		  <ExcludeAssets>runtime</ExcludeAssets>
 	  </PackageReference>
-	  <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" >
+	  <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1">
 		  <Private>false</Private>
 		  <ExcludeAssets>runtime</ExcludeAssets>
 	  </PackageReference>
